Now is possible to add inactive / bad aquapoints;
Aquapoint dashboard modifications
General dashboard modifications
Now is possible to send a image when reporting a aquapoint
Updated SQL files to support this new systems
Users can now see their pending aquapoint requests on map

+++ Front_End/AquaPoint/src/components/AquapointDetailsModal.vue
@@ -232,6 +232,11 @@ async function ChangeFavoriteState() {
         return
     }
 
+    if (props.aquapoint.isPending == 1) {
+        toast.error('O bebedouro ainda está pendente de aprovação.')
+        return
+    }
+
     if (isFavorite.value) {
         await favoriteService.delete({ user_id: Auth.user.id, point_id: props.aquapoint.id })
         toast.info('Bebedouro removido dos favoritos.')
@@ -269,6 +274,11 @@ async function SubmitReview() {
         return
     }
 
+    if (props.aquapoint.isPending == 1) {
+        toast.error('O bebedouro ainda está pendente de aprovação.')
+        return
+    }
+
     if (reviewText.value.trim() === '') {
         toast.error('Por favor, escreva um comentário antes de submeter a sua opinião.')
         return
@@ -299,6 +309,11 @@ async function SubmitReview() {
 function ReportProblem() {
     console.log('Report Problem clicked')
 
+    if (props.aquapoint.isPending == 1) {
+        toast.error('O bebedouro ainda está pendente de aprovação.')
+        return
+    }
+    
     showReportProblemModal.value = true
 }
 
@@ -431,6 +446,15 @@ async function ApplyCommentChanges(review){
     left: 25%;
     border-radius: 8px;
     object-fit: contain;
+}
+
+.image {
+    position: absolute;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 14rem;
+    background-color: red;
 }
 
 .modal-box>img {
